requirements.txt
blinker==1.8.2
click==8.1.8
coloredlogs==15.0.1
flatbuffers==25.2.10
humanfriendly==10.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
numpy<2
onnxruntime-gpu==1.16.3
opencv-python-headless
# opencv-python==4.9.0.80
packaging==25.0
protobuf==5.29.5
setuptools==75.1.0
sympy==1.13.3
tqdm==4.67.1
Werkzeug==3.0.6
wheel==0.44.0
zipp==3.20.2
